4_Automacao_IA/requirements.txt
python-dotenv
feedparser
requests
openai
firecrawl-py==1.10.0
